Multiply Hit Rate by 100 to calculate expected percentage value (#28)
* Prevent divide by zero error

When there were no hits or misses against a node, the result being returned was NaN as a result of a divide by zero issue. This approach converts the NaN result to a zero.

* Multiply Hit Rate ratio by 100 to calculate percentage

The Hit Rate metric unit is Percent but we are sending a ratio. We need to multiply the ratio by 100 to get a percentage so monitoring rules and metrics make sense.

Resolves #27

Co-authored-by: Erik Ruggles <erikr@tempworks.com>